fix(app): Material 泄漏治理与设计系统语义对齐
- 全量将透明/占位色改为 AppColors 语义 token，补齐 AppSpacing 圆角与欢迎页几何常量
- 欢迎页改用 AppScaffold；助理工具栏/再生选项/发现宫格等改用 token
- 沉浸式作品浏览器测试改为查找 CupertinoIcons.ellipsis（与实现一致）
- 特性树与 iOS native surface 门禁清单同步更新

Made-with: Cursor

+++ quwoquan_app/lib/assistant/tool/impl/search/search_tool.dart
@@ -647,8 +647,8 @@ class SearchTool implements AssistantTool {
     score += referenceCount > 0 ? 0.25 : 0.0;
     score += includesWeb && referenceCount > 0 ? 0.1 : 0.0;
     score += includesInternal && hitCount > 0 ? 0.15 : 0.0;
-    score += (hitCount.clamp(0, 6) as int) * 0.04;
-    score += (referenceCount.clamp(0, 6) as int) * 0.04;
+    score += (hitCount.clamp(0, 6)) * 0.04;
+    score += (referenceCount.clamp(0, 6)) * 0.04;
     score += queryCount > 1 ? 0.1 : 0.05;
     if (score > 0.95) {
       return 0.95;
